move process dynamic links to separate widget
